server/requirements.txt
django==4.1.5
djangorestframework==3.14.0
psycopg2==2.9.5
drf_yasg==1.21.4
djoser==2.1.0
django-cors-headers==3.13.0
